Move tr_prefix_wildcard_match to tr_constraint.c
We need tr_prefix_wildcard_match for merge_constraints and for
tr_filter.c.  Export it from libtr_tid even though it's in a private
header.  It's not part of the public API but is part of the library so
tr_filter can import it.

Also, fix bug; all strings were treated as wildcards.