Reload configuration dynamically when files change.
Detect a change to the configuration files and reload when this occurs.
Any change, deletion, or addition of configuration files in the
config directory will result in reloading the configuration. If a
valid configuration results, switch to this, otherwise continue using
the old configuration. Allows a few seconds for changes to settle
before attempting to reload the configuration.