Use our hostname/TID port when sending a request, not our next_hop
Before this, we set the next_hop to ourselves for local routes, then
simply forwarded the next_hop to our peers in update messages. That is
incorrect - we need to fill in our own hostname/TID port every time, not
send the next_hop we forward to.

Also fixes a few port name / signed int changes that really belonged in
the previous commit.