Make the TID client to interpret the content of aaa_server_addr
This allows "aaa_server_addr" to be a hostname, and IPv4 or an
IPv6 address.

Incidentally, this also allows hostnames to be used rather than IP
addresses. This can be more convenient in cases where the server does not
want to impose IPv4 or IPv6 and rather let the client decide.